Cart Items Component created
